+++ src/models/appointment.js
@@ -31,7 +31,7 @@ const appointmentSchema = new mongoose.Schema(
   },
   {
     timestamps: true,
-  },
+  }
 );
 
 // Export the model