+++ src/models/mealplan.js
@@ -35,7 +35,7 @@ const mealPlanSchema = new mongoose.Schema(
   },
   {
     timestamps: true,
-  },
+  }
 );
 
 // Export the model